requirements.txt
streamlit
requests
numpy
pillow
